+++ bindings/nodejs/lib/types/utils/bridge/index.ts
@@ -20,6 +20,7 @@ import type {
     __IsAddressValidMethod__,
     __HashTransactionEssenceMethod__,
     __VerifyEd25519SignatureMethod__,
+    __VerifySecp256k1EcdsaSignatureMethod__,
     __VerifyMnemonicMethod__,
     __FaucetMethod__,
 } from './utils';
@@ -46,5 +47,6 @@ export type __UtilsMethods__ =
     | __IsAddressValidMethod__
     | __HashTransactionEssenceMethod__
     | __VerifyEd25519SignatureMethod__
+    | __VerifySecp256k1EcdsaSignatureMethod__
     | __VerifyMnemonicMethod__
     | __FaucetMethod__;