+++ sdk/tests/client/node_api.rs
@@ -6,7 +6,7 @@
 use iota_sdk::{
     client::{
         api::GetAddressesOptions, bech32_to_hex, node_api::indexer::query_parameters::QueryParameter,
-        request_funds_from_faucet, secret::SecretManager, Client,
+        request_funds_from_faucet, secret::SecretManager, Client, Error,
     },
     types::block::{
         address::ToBech32Ext,
@@ -419,4 +419,26 @@ async fn test_mqtt() {
         }
     }
     client.subscriber().disconnect().await.unwrap();
+}
+
+#[ignore]
+#[tokio::test]
+async fn test_plugin_fetch() {
+    let c = setup_client_with_node_health_ignored().await;
+
+    let nft_id = "0x0000000000000000000000000000000000000000000000000000000000000000"
+        .try_into()
+        .unwrap();
+
+    let route = format!("outputs/nft/{nft_id}");
+
+    let nft_1 = c
+        .plugin_fetch::<OutputId>("api/indexer/v1/", &"GET", &route, vec![], None)
+        .await;
+    let nft_2 = c.nft_output_id(nft_id).await;
+
+    // Actually this returns Err(Error::Node(crate::client::node_api::error::Error::NotFound(e)))
+    if let (Ok(id_1), Ok(id_2)) = (&nft_1, &nft_2) {
+        assert_eq!(id_1, id_2);
+    }
 }